requirements.txt
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
numpy
tensorboardX
ffmpeg-python
opencv-python